refactor: Profile - Modals 디렉토리 추가 및 분류
