+++ src/profile_page/components/Modals/EditCourseModal.tsx
@@ -1,30 +1,36 @@
 import { useRecoilState, useRecoilValue } from "recoil";
-import { userId, userName } from "../../recoil/userInfo";
-import style from "../Profile.module.css";
+import { userId, userName } from "../../../recoil/userInfo";
+import style from "../../Profile.module.css";
 import { useState } from "react";
 import { CircleAlert } from "lucide-react";
-import { api } from "../../utils/api";
+import { api } from "../../../utils/api";
 
 interface Props {
   onClose: () => void;
+  courseId: string;
+  courseTitle: string;
+  onUpdateTitle: (newTitle: string) => void;
 }
 
-const EditNameModal = ({ onClose }: Props) => {
-  const [userNameState, setUserNameState] = useRecoilState(userName);
-  const [newName, setNewName] = useState(userNameState); // 로컬 상태로 복사
+const EditCourseModal = ({
+  onClose,
+  courseId,
+  courseTitle,
+  onUpdateTitle,
+}: Props) => {
+  const [newTitle, setNewTitle] = useState<string>(courseTitle); // 로컬 상태로 복사
   const token = localStorage.getItem("accessToken");
-  const userIdState = useRecoilValue(userId);
 
   const handleSubmit = async () => {
-    if (newName.trim()) {
-      const nameTrim = newName.trim();
+    if (newTitle.trim()) {
+      const titleTrim = newTitle.trim();
 
       try {
         await api.put(
-          "/user",
+          "/course/title",
           {
-            id: userIdState,
-            name: nameTrim,
+            courseId: courseId,
+            title: titleTrim,
           },
           {
             headers: {
@@ -32,12 +38,10 @@ const EditNameModal = ({ onClose }: Props) => {
             },
           },
         );
-
-        setUserNameState(nameTrim);
-        localStorage.setItem("userName", nameTrim);
+        onUpdateTitle(titleTrim);
         onClose();
       } catch (error) {
-        alert("name error");
+        alert("title error");
         console.log(error);
       }
     }
@@ -54,14 +58,14 @@ const EditNameModal = ({ onClose }: Props) => {
           }}
         >
           <CircleAlert color="var(--color-main)" size={26} />
-          <div className={style.title}>&nbsp;프로필 수정</div>
+          <div className={style.title}>&nbsp;코스 이름 수정</div>
         </div>
         <div>
-          <p className={style.nameTitle}>닉네임</p>
+          <p className={style.nameTitle}>코스 이름</p>
           <input
             type="text"
-            value={newName}
-            onChange={(e) => setNewName(e.target.value)}
+            value={newTitle}
+            onChange={(e) => setNewTitle(e.target.value)}
             className={style.inputName}
           />
         </div>
@@ -78,4 +82,4 @@ const EditNameModal = ({ onClose }: Props) => {
   );
 };
 
-export default EditNameModal;
+export default EditCourseModal;
